Tasks/bug fixing

+++ src/app/components/modals/chat-information/chanel-settings/chanel-settings.component.ts
@@ -1,11 +1,15 @@
 import { Component, OnInit, Input } from '@angular/core';
-
-import { SocketMessageModel } from '../../../../models/socket.message.model';
+import { MatDialog } from '@angular/material';
 
 import { ChatService} from '../../../../services/chat/chat.service';
 import { SocketsService } from '../../../../services/sockets/sockets.service';
 import { AuthService } from '../../../../services/auth/auth.service';
 import { RequestsService } from '../../../../services/requests/requests.service';
+
+import { MessageModalComponent } from '../../message-modal/message-modal.component';
+
+import { MessageModalModel } from '../../message-modal/message-modal.model';
+import { SocketMessageModel } from '../../../../models/socket.message.model';
 
 @Component({
   selector: 'app-chanel-settings',
@@ -19,6 +23,7 @@ export class ChanelSettingsComponent implements OnInit {
   @Input() chatUsers: any;
 
   constructor(
+    private dialog: MatDialog,
     private authService: AuthService,
     private chatService: ChatService,
     private api: RequestsService,
@@ -35,8 +40,17 @@ export class ChanelSettingsComponent implements OnInit {
   }
 
   public deleteChannel(): void {
-    this.api.delete({url: `/chats/${this.chatService.activeChat._id}`})
-      .subscribe(() => console.log('Success'));
+    const dialogRef = this.dialog.open(MessageModalComponent, {
+      width: '450px',
+      data: new MessageModalModel('Delete channel?', true, true)
+    });
+    dialogRef.afterClosed()
+      .subscribe(res  => {
+        if (res) {
+          this.api.delete({url: `/chats/${this.chatService.activeChat._id}`})
+            .subscribe(() => console.log('Success'));
+        }
+      });
   }
 
 }
